Folder restructe and code cleanup
