+++ src/navigation/DrawerNav/components/ActiveRecall.jsx
@@ -10,15 +10,12 @@ import {
 } from "react-native";
 import React from 'react';
 import { Button, Appbar } from "react-native-paper";
-import { colors } from "../utils/colors";
-import Header from "../components/Header";
-import TopBar from "../components/TopBar";
+
 
 const ActiveRecall = () => {
     return (
-        <SafeAreaView>
-          <TopBar/>
-        </SafeAreaView>
+        <View>
+        </View>
       );
 }
