+++ src/navigation/TopTabNav/components/Done.jsx
@@ -1,11 +1,11 @@
 import { View, Text, StyleSheet } from 'react-native';
 import React from 'react';
-import { colors } from '../utils/colors';
+import { colors } from '../../../utils/colors';
 
 const Done = () => {
     return (
         <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-          <Text style={styles.label}>Done...</Text>
+          <Text style={styles.label}>Donessd...</Text>
         </View>
       );
 }
